requirements.txt
habitat.dist==info
torch==1.9.0
numpy==1.21.2
torchvision==0.10.0
gym==0.19.0
seaborn==0.11.2
gym_minigrid==1.0.2
pandas==1.3.2
tqdm==4.62.1
habitat_sim==0.2.1
matplotlib==3.4.3
submitit==1.3.3
numpy_quaternion==2021.7.13.7.29.27
GitPython==3.1.24
pickle5==0.0.11
Pillow==8.3.2
quaternion==3.5.2.post4
git+https://github.com/facebookresearch/detectron2.git
git+https://github.com/openai/CLIP.git
timm=0.5.4
